Add serial input capabilities to board-view
This adds the ability to send keyboard input from board-view to use the
CircuitPython REPL with the connected device. It also changes the html
and styling to make the view's style fit the original look of the
package and whichever theme is currently set.

With this functionality, this package provides an equivalent set of
features to those that are included in the mu-editor, which is the
editor Adafruit recommends as it is 'designed for' CircuitPython. Now
Atom could be just as useful when using CircuitPython.

I want to add that I converted a good portion of the Serial
communications code from the mu-editor source code. Job well done by
them for implementing that, its a pretty cool feature/implementation.

This should work on Mac and a future commit will add full support for
Windows.