Add full support for serial input on Windows
Add code to handle the keystrokes that are used on Windows but not Mac
(i.e. del, home, end, control).

Also add a new serial parser that will wait for a slight delay before
presenting the serial input data to the board-view. I noticed when
running on my Windows machine that serial output looked a bit jumpy and
the cursor also looked like it was jumping when it was printing out
text. It turned out to be because the serial data was being presented as
separate chunks. I'm not sure why this happened here and not on my Mac,
but either way now the parser will send all the data that it receives at
'virtually' the same time, as one chunk to board-view.