Truncate output when flooded
If the microcontroller is sending data too rapidly for us to keep up,
the output is truncated. This prevents the printing from falling too far
behind the microcontroller's output and becoming sluggish or
unresponsive. This feature can be turned on or off through the settings.

This feature is mainly for when there is a loop printing out text with
no delay between each print call.