Downgrade serialport to v9.0.3
This will downgrade serialport and @serialport/bindings to 9.0.3. After
testing through a bunch of the different serialport versions I found
that this is the newest version that doesn't cause Atom to crash.
However, @serialport/bindings has to be locked at 9.0.3 in order for
this to work. I'm assuming this is why some might have been experiencing
problems, while others were not.